@@ -2,30 +2,24 @@ source 'https://rubygems.org'
 
 # The core framework
 # https://github.com/rails/rails
-gem 'rails', '4.2.11.1'
+gem 'rails', '5.0.7.2'
 
 gem 'actionpack-action_caching', '~>1.2.1'
 gem 'actionpack-page_caching', '~>1.1.1'
 gem 'actionpack-xml_parser', '~>1.0.1'
 gem 'actionview-encoded_mail_to', '~>1.0.5'
 gem 'activerecord-session_store', '~>1.1.0'
-gem 'protected_attributes', '~> 1.0.8'
 gem 'rails-observers', '~>0.1.2'
-gem 'rails-perftest', '~>0.0.5'
 gem 'responders', '~> 2.0'
 
 gem 'rake', '~> 12.3.3'
 
-# ETS is temporarily maintaining its own fork of Rack to support the SameSite=None cookie property.
-gem 'rack', '~> 1.6.13', git: 'https://github.com/ets-berkeley-edu/rack.git', branch: 'ets-1-6-samesite'
+gem 'rack', '~> 2.2.3'
 
-gem 'activerecord-jdbc-adapter', '~> 1.3.16'
+gem 'activerecord-jdbc-adapter', '~> 50.7'
 
 # Postgresql adapter
-gem 'activerecord-jdbcpostgresql-adapter', '~> 1.3.16'
-
-# H2 adapter
-gem 'activerecord-jdbch2-adapter', '~> 1.3.16'
+gem 'activerecord-jdbcpostgresql-adapter', '~> 50.7'
 
 # A JSON implementation as a Ruby extension in C
 # http://flori.github.com/json/
