@@ -1,14 +1,10 @@
 module CanvasCsv
   # Provides object used to store synchronization states between campus systems and Canvas
-  class Synchronization < ActiveRecord::Base
+  class Synchronization < ApplicationRecord
     include ActiveRecordHelper
     include ClassLogger
 
     self.table_name = 'canvas_synchronization'
-    attr_accessible :last_enrollment_sync
-    attr_accessible :last_guest_user_sync
-    attr_accessible :last_instructor_sync
-    attr_accessible :latest_term_enrollment_csv_set
 
     # Returns single record used to store synchronization timestamp(s)
     def self.get
