@@ -1,8 +1,7 @@
 module Oec
-  class CourseCode < ActiveRecord::Base
+  class CourseCode < ApplicationRecord
 
     self.table_name = 'oec_course_codes'
-    attr_accessible :dept_name, :catalog_id, :dept_code, :include_in_oec
 
     def to_h
       Hash[[:dept_name, :catalog_id, :dept_code, :include_in_oec].collect {|m| [m, send(m)]}]
