@@ -1,7 +1,6 @@
-# Load the rails application
-require File.expand_path('../application', __FILE__)
-
+# Load the Rails application.
+require_relative 'application'
 require 'openssl'
 
-# Initialize the rails application
-Calcentral::Application.initialize!
+# Initialize the Rails application.
+Rails.application.initialize!
